Added skript to remove category URLs.
When reorganizing the category tree the category URLs remain the same. Only
new categories will get a newly assigned URL. This may result in duplicate
URL, which show up in the dashboard. The script allows to reset the category
URLs. The URLs will then be newly assigned during the first access.