Reorganisation completed.
The center-south area (mainly used for plugins) is now part of the
main center area and the east region is part of the inner center
area. That causes the plugin area to have more available with.

Additionally a dblclick listener on the south-center region allows
to collapse the west region and to resize the south-center region up to
600 px (and back).