Removed passing params by reference.
There seems to be a bug or design error in the method call_user_func
of PHP. Before PHP 5.3 the parameters may have been passed by reference
which seems to be deprecated in the meantime. The prototyping in the
referenced method does not matter. Either way (taking or not as reference)
will result in cloning the paramter.

The methods beginning with _extraFieldAtts have been changed to take
a clone of the param and to return the changed value.