Changed behaviour of category tree.
The category tree now does not load the category content into the main
area, rather than just expand the node, if it is expandable. If the node
is already expanded or the if the node is not expandable (i.e. it does
not contain any children) the category content is loaded into the main
area.